sumaDe puntos fotos

+++ app/src/main/java/com/example/snapking/modelo/Jugador.kt
@@ -1,7 +1,7 @@
 package com.example.snapking.modelo
 
 
-data class Jugador(var id: String, var ready:Boolean, var punto:Int,var etapa:Etapa) {
-    constructor() : this("", true,0,Etapa.LOBBY) {
+data class Jugador(var id: String, var ready:Boolean, var punto:Float,var etapa:Etapa) {
+    constructor() : this("", true,0f,Etapa.LOBBY) {
     }
 }
